Add Features window
Adds a window where the developer can choose which Windows features he/she
wants to integrate in the app.

These settings are saved to a file that is versioned, so everyone gets the
same features.

A build in Unity is needed in order for changes in the settings to take
effect in the Windows solution.