Remove ExceptionLogger monobehaviour
Removes the ExceptionLogger class and changes IntegrationManager to be
a static class that should initialize all the stuff the today is using
the Awake method of a MonoBehaviour to make its initialization.

IntegrationManager.Init() should be called on the Windows side as soon
as Unity is ready to run its things.

Made a scene change to remove ExceptionLogger from the game object it
was attached to.